fix(world): updateEnergy uses float-then-truncate drain math (player-core-2)
TS Player.updateEnergy (Player.ts:690-693) keeps weightKg as float
(`this.runweight / 1000`), clamps the float to [0, 64], then truncates
only the final loss expression via `| 0`. goscape's pre-fix code did
`weightKg := p.runweight / 1000` as int division, dropping the kg
fraction BEFORE the `67*weightKg/64` math, so any partial-kg
encumbrance (the common in-game case) under-drained run energy.

With runweight=32500 (32.5 kg):
  TS:  (67 + 67*32.5/64) | 0 = (67 + 34.023) | 0 = 101 drain
  bug: 67 + 67*32/64 (int div) = 67 + 33 = 100 drain
After the float port: both yield 101.

Add TestUpdateEnergy_DrainFractionalKgUsesFloatMath pinning runweight=
32500 → 9899 (RED 9900 → GREEN 9899). The five existing drain tests
all use integer-kg values (0, 64000, 200000, -1000) where the bug
silently agreed with TS — none caught the divergence, hence the audit
row. They remain green under the float port (negative clamps to 0
float, ≥64kg clamps to 64.0 float — both yield identical loss).

Also realign the npc.go struct-literal padding after the npc-ai-3
`nextPatrolTick: -1` insertion — gofmt-required follow-up.

+++ modules/world/player_run.go
@@ -35,9 +35,15 @@ func (p *Player) updateEnergy() {
 		recovered := agility/9 + 8
 		p.runenergy = min(p.runenergy+recovered, 10000)
 	} else {
-		weightKg := p.runweight / 1000
+		// player-core-2: TS Player.ts:690-693 keeps weightKg as float
+		// (`this.runweight / 1000`) and only truncates the final loss
+		// via `| 0`, so a partial-kg encumbrance contributes a
+		// fractional drain that rounds away properly at the end. Int
+		// division on runweight here would drop the fraction BEFORE
+		// the 67*weightKg/64 math and systematically under-drain.
+		weightKg := float64(p.runweight) / 1000
 		clampWeight := max(min(weightKg, 64), 0)
-		loss := 67 + 67*clampWeight/64
+		loss := int(67 + 67*clampWeight/64)
 		p.runenergy = max(p.runenergy-loss, 0)
 	}
 	if p.runenergy == 0 {
